+++ project/history.go
@@ -58,6 +58,7 @@ var History relic.ImmutableHistory = relic.NewHistory("Hyperledger Burrow", "htt
 
 ### Added
 - [Kernel] Added announce message for startup and shutdown including version, key address, and other useful metadata
+- [EVM] Attempt to provide REVERT reason where possible
 `,
 		"0.24.1 - 2019-02-28",
 		`### Changed